# backend_monbrobroli-backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "L5 OpenApi",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "darius@matulionis.lt"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0",
        "x-logo": {
            "url": "https://via.placeholder.com/190x90.png?text=L5-Swagger"
        }
    },
    "paths": {
        "api/create_categorie-abonnement": {
            "post": {
                "tags": [
                    "Abonnement."
                ],
                "summary": "Creation d'abonnement",
                "description": "Creation d'abonnement",
                "operationId": "f406dfcd6e5525a101c47099b44325b9",
                "requestBody": {
                    "description": "création de catégorie abonnement",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "libelle": {
                                        "type": "string",
                                        "example": "Etudiant"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "...."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "api/getCategorie": {
            "get": {
                "tags": [
                    "Abonnement."
                ],
                "summary": "Voir toutes les catégories",
                "description": "Voir toutes les catégories",
                "operationId": "Voir toutes les catégories",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
...
        "api/deleteCategorie/{id}": {
            "delete": {
                "tags": [
                    "Abonnement."
                ],
                "summary": "Supprimer une catégorie",
                "description": "Supprimer une catégorie",
                "operationId": "deleteCategorie",
                "parameters": [
                    {
                        "name": "Id",
                        "in": "path",
                        "description": "Id de l'emploi du temps.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid schedule supplied"
                    },
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Catégorie supprimée avec succes"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "route not found"
                    }
                }
            }
        },
        "api/editCategorie/{id}": {
            "put": {
                "tags": [
                    "Abonnement."
                ],
                "summary": "Modifier la catégorie.",
                "operationId": "Modifier la catégorie",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id de l'emploi du temps.",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Modifier la catégorie",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "libelle": {
                                        "type": "string",
                                        "example": "Etudiant"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "400": {
                        "description": "Invalid ID supplied"
                    },
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "405": {
                        "description": "Mauvaise méthode"
                    }
                }
            }
        },
        "api/create-abonnement": {
            "post": {
                "tags": [
                    "Abonnement."
                ],
                "summary": "Creation d'abonnement",
                "description": "Creation d'abonnement",
                "operationId": "30190b2c32906251edbb07f8c1778494",
                "requestBody": {
                    "description": "création de catégorie abonnement",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "libelle": {
                                        "type": "string",
                                        "example": "Etudiant"
                                    },
                                    "prix": {
                                        "type": "number",
                                        "example": "2000"
                                    },
                                    "periode": {
                                        "type": "number",
                                        "example": "1 ou 2 ans ( nb:ne pas écrire le ans)"
                                    },
                                    "categorie_id": {
                                        "type": "number",
                                        "example": "4"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "string",
                                            "example": "[{}]"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "api/admin/deleteAbonnement/{id}": {
            "delete": {
                "tags": [
                    "Abonnement."
                ],
                "summary": "Supprimer un abonnement",
                "description": "Supprimer un abonnement",
                "operationId": "deleteAbonnement",
                "parameters": [
                    {
                        "name": "Id",
                        "in": "path",
                        "description": "Id de l'abonnement.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid schedule supplied"
                    },
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Catégorie supprimée avec succes"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "route not found"
                    }
                }
            }
        },
        "api/getAbonnement": {
            "get": {
                "tags": [
                    "Abonnement."
                ],
                "summary": "Voir touts les abonnements",
                "description": "Voir touts les abonnements",
                "operationId": "Voir touts les abonnements",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
...
    },
...
}